+++ Security/UserVoter.php
@@ -11,23 +11,87 @@ namespace c975L\UserBundle\Security;
 
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
+use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class UserVoter extends Voter
 {
+    /**
+     * Stores ContainerInterface
+     * @var ContainerInterface
+     */
     private $container;
+
+    /**
+     * Stores AccessDecisionManagerInterface
+     * @var AccessDecisionManagerInterface
+     */
     private $decisionManager;
+
+    /**
+     * The role needed to be allowed access (defined in config)
+     * @var string
+     */
     private $roleNeeded;
 
-    public const CHANGE_PASSWORD = 'change-password';
-    public const DASHBOARD = 'dashboard';
-    public const DELETE = 'delete';
-    public const DISPLAY = 'display';
-    public const EXPORT = 'export';
-    public const HELP = 'help';
-    public const MODIFY = 'modify';
-    public const PUBLIC_PROFILE = 'public-profile';
-    public const RESET_PASSWORD = 'reset-password';
+    /**
+     * Used for access to change-password
+     * @var string
+     */
+    public const CHANGE_PASSWORD = 'c975LUser-change-password';
 
+    /**
+     * Used for access to dashboard
+     * @var string
+     */
+    public const DASHBOARD = 'c975LUser-dashboard';
+
+    /**
+     * Used for access to delete
+     * @var string
+     */
+    public const DELETE = 'c975LUser-delete';
+
+    /**
+     * Used for access to display
+     * @var string
+     */
+    public const DISPLAY = 'c975LUser-display';
+
+    /**
+     * Used for access to export
+     * @var string
+     */
+    public const EXPORT = 'c975LUser-export';
+
+    /**
+     * Used for access to help
+     * @var string
+     */
+    public const HELP = 'c975LUser-help';
+
+    /**
+     * Used for access to modify
+     * @var string
+     */
+    public const MODIFY = 'c975LUser-modify';
+
+    /**
+     * Used for access to public-profile
+     * @var string
+     */
+    public const PUBLIC_PROFILE = 'c975LUser-public-profile';
+
+    /**
+     * Used for access to reset-password
+     * @var string
+     */
+    public const RESET_PASSWORD = 'c975LUser-reset-password';
+
+    /**
+     * Contains all the available attributes to check with in supports()
+     * @var array
+     */
     private const ATTRIBUTES = array(
         self::CHANGE_PASSWORD,
         self::DASHBOARD,
@@ -41,8 +105,8 @@ class UserVoter extends Voter
     );
 
     public function __construct(
-        \Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface $decisionManager,
-        \Symfony\Component\DependencyInjection\ContainerInterface $container,
+        AccessDecisionManagerInterface $decisionManager,
+        ContainerInterface $container,
         string $roleNeeded
     )
     {
@@ -51,6 +115,10 @@ class UserVoter extends Voter
         $this->roleNeeded = $roleNeeded;
     }
 
+    /**
+     * Checks if attribute and subject are supported
+     * @return bool
+     */
     protected function supports($attribute, $subject)
     {
         if (false !== $subject) {
@@ -60,6 +128,11 @@ class UserVoter extends Voter
         return in_array($attribute, self::ATTRIBUTES);
     }
 
+    /**
+     * Votes if access is granted
+     * @return bool
+     * @throws \LogicException
+     */
     protected function voteOnAttribute($attribute, $subject, TokenInterface $token)
     {
         //Defines access rights
@@ -85,19 +158,28 @@ class UserVoter extends Voter
         throw new \LogicException('Invalid attribute: ' . $attribute);
     }
 
-    //Checks if public profile is allowaed by wedsite
+    /**
+     * Checks if public profile is allowed by website
+     * @return bool
+     */
     public function isAllowedPublicProfile()
     {
         return $this->container->getParameter('c975_l_user.publicProfile');
     }
 
-    //Checks if user is owner or has admin rights
+    /**
+     * Checks if user is owner or has admin rights
+     * @return bool
+     */
     private function isOwner($token, $user)
     {
         return $this->isAdmin($token) ? true : $user->getId() === $token->getUser()->getId();
     }
 
-    //Checks if user has admin rights
+    /**
+     * Checks if user has admin rights
+     * @return bool
+     */
     private function isAdmin($token)
     {
         return $this->decisionManager->decide($token, array($this->roleNeeded));